BEZIER CURVES
* actual fruit animation
* toggle for random fruits (doesn't work yet)
* textbox for custom objects (doesn't work yet)
 - not user-submitted stuff, just stuff that doesn't really fit in the main dropdown menu
* some stuff i don't remember, but it's definitely stuff that's working towards making everything work.

miss anims don't work yet. at all. the objects just go flying. also the slice animation needs to be finished for both sides, and a bool needs to be added for which side to play (will be easy to do) also when you randomly slice, which is determined completely randomly (thank god)

gura-nyuuu :3 (i'm sorry)
* added a single sfx