More Animation work and Modernized BG Event
Added animations to move the whole doll on a just/missed input. Fixed bugs with the default hand object not properly playing its animations due to its names previously being changed. Modernized the BG Colors event to have the whole easing dropdown in place of the "Fade?" toggle. Combined the spotlight recoloring section of the BG Colors event with the Hand Colors event to make the Object Colors event.